Tests/LibWeb/Text/expected/details-inside-slot-opened.txt

Height before opening: 18px
Height after opening: 36px
